+++ Dominio/Modelo/ClsVentaDetallDom.cs
@@ -14,8 +14,8 @@ namespace Dominio.Modelo
 
         public int ProductoID { get; set; }
 
-        public double Cantidad { get; set; }
+        public int Cantidad { get; set; }
 
-        public double Precio { get; set; }
+        public int Precio { get; set; }
     }
 }
